+++ src/main/kotlin/no/nav/helsearbeidsgiver/domene/inntektsmelding/AarsakInnsending.kt
@@ -3,6 +3,7 @@ package no.nav.helsearbeidsgiver.domene.inntektsmelding
 import kotlinx.serialization.Serializable
 import kotlin.String
 
+@Deprecated("Bruk 'v1.AarsakInnsending' istedenfor.")
 @Serializable
 enum class AarsakInnsending(val value: String) {
     NY("Ny"),
