+++ src/main/kotlin/no/nav/helsearbeidsgiver/domene/inntektsmelding/Innsending.kt
@@ -4,9 +4,11 @@ package no.nav.helsearbeidsgiver.domene.inntektsmelding
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
+import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.Periode
 import no.nav.helsearbeidsgiver.utils.json.serializer.LocalDateSerializer
 import java.time.LocalDate
 
+@Deprecated("Bruk 'v1.SkjemaInntektsmelding' istedenfor.")
 @Serializable
 data class Innsending(
     val orgnrUnderenhet: String,
