+++ src/main/kotlin/no/nav/helsearbeidsgiver/domene/inntektsmelding/Kvittering.kt
@@ -4,11 +4,13 @@ package no.nav.helsearbeidsgiver.domene.inntektsmelding
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
+import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.Periode
 import no.nav.helsearbeidsgiver.utils.json.serializer.LocalDateSerializer
 import no.nav.helsearbeidsgiver.utils.json.serializer.OffsetDateTimeSerializer
 import java.time.LocalDate
 import java.time.OffsetDateTime
 
+@Deprecated("Bruk 'v1.Kvittering' istedenfor.")
 @Serializable
 data class Kvittering(
     val kvitteringDokument: KvitteringSimba? = null,
@@ -28,16 +30,17 @@ data class KvitteringSimba(
     val arbeidsgiverperioder: List<Periode>,
     val bestemmendeFraværsdag: LocalDate,
     val fraværsperioder: List<Periode>,
-    val inntekt: InntektDeprecated,
+    val inntekt: Inntekt,
     val fullLønnIArbeidsgiverPerioden: FullLoennIArbeidsgiverPerioden? = null,
-    val refusjon: RefusjonDeprecated,
-    val naturalytelser: List<NaturalytelseDeprecated>? = null,
+    val refusjon: Refusjon,
+    val naturalytelser: List<Naturalytelse>? = null,
     val årsakInnsending: AarsakInnsending,
     val bekreftOpplysninger: Boolean,
     val tidspunkt: OffsetDateTime,
     val forespurtData: List<String>? = null,
 )
 
+@Deprecated("Bruk 'v1.KvitteringEkstern' istedenfor.")
 @Serializable
 data class KvitteringEkstern(
     val avsenderSystem: String,
