+++ src/main/kotlin/no/nav/helsearbeidsgiver/domene/inntektsmelding/v1/AarsakInnsending.kt
@@ -1,0 +1,9 @@
+package no.nav.helsearbeidsgiver.domene.inntektsmelding.v1
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class AarsakInnsending {
+    Ny,
+    Endring,
+}
