+++ src/main/kotlin/no/nav/helsearbeidsgiver/domene/inntektsmelding/v1/Inntekt.kt
@@ -1,6 +1,6 @@
 @file:UseSerializers(LocalDateSerializer::class)
 
-package no.nav.helsearbeidsgiver.domene.inntektsmelding
+package no.nav.helsearbeidsgiver.domene.inntektsmelding.v1
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
@@ -8,13 +8,9 @@ import no.nav.helsearbeidsgiver.utils.json.serializer.LocalDateSerializer
 import java.time.LocalDate
 
 @Serializable
-data class Periode(
-    val fom: LocalDate,
-    val tom: LocalDate,
+data class Inntekt(
+    val beloep: Double,
+    val inntektsdato: LocalDate,
+    val naturalytelser: List<Naturalytelse>,
+    val endringAarsak: InntektEndringAarsak? = null,
 )
-
-infix fun LocalDate.til(tom: LocalDate): Periode =
-    Periode(
-        fom = this,
-        tom = tom,
-    )
