+++ src/main/kotlin/no/nav/helsearbeidsgiver/domene/inntektsmelding/v1/Inntektsmelding.kt
@@ -1,0 +1,23 @@
+@file:UseSerializers(OffsetDateTimeSerializer::class, UuidSerializer::class)
+
+package no.nav.helsearbeidsgiver.domene.inntektsmelding.v1
+
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.UseSerializers
+import no.nav.helsearbeidsgiver.utils.json.serializer.OffsetDateTimeSerializer
+import no.nav.helsearbeidsgiver.utils.json.serializer.UuidSerializer
+import java.time.OffsetDateTime
+import java.util.UUID
+
+@Serializable
+data class Inntektsmelding(
+    val id: UUID,
+    val sykmeldt: Sykmeldt,
+    val avsender: Avsender,
+    val sykmeldingsperioder: List<Periode>,
+    val agp: Arbeidsgiverperiode?,
+    val inntekt: Inntekt,
+    val refusjon: Refusjon,
+    val aarsakInnsending: AarsakInnsending,
+    val mottatt: OffsetDateTime,
+)
