+++ src/main/kotlin/no/nav/helsearbeidsgiver/domene/inntektsmelding/v1/Naturalytelse.kt
@@ -1,21 +1,19 @@
 @file:UseSerializers(LocalDateSerializer::class)
 
-package no.nav.helsearbeidsgiver.domene.inntektsmelding
+package no.nav.helsearbeidsgiver.domene.inntektsmelding.v1
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
 import no.nav.helsearbeidsgiver.utils.json.serializer.LocalDateSerializer
 import java.time.LocalDate
 
-@Deprecated("Bruk 'v1.Naturalytelse' istedenfor.")
 @Serializable
 data class Naturalytelse(
     val naturalytelse: NaturalytelseKode,
-    val dato: LocalDate,
-    val beløp: Double,
+    val verdiBeloep: Double,
+    val sluttdato: LocalDate,
 )
 
-@Deprecated("Bruk 'v1.NaturalytelseKode' istedenfor.")
 @Serializable
 enum class NaturalytelseKode {
     AKSJERGRUNNFONDSBEVISTILUNDERKURS,
