+++ src/main/kotlin/no/nav/helsearbeidsgiver/domene/inntektsmelding/v1/skjema/SkjemaInntektsmelding.kt
@@ -1,0 +1,19 @@
+package no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.skjema
+
+import kotlinx.serialization.Serializable
+import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.AarsakInnsending
+import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.Arbeidsgiverperiode
+import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.Inntekt
+import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.Periode
+import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.Refusjon
+
+@Serializable
+data class SkjemaInntektsmelding(
+    val sykmeldtFnr: String,
+    val avsender: SkjemaAvsender,
+    val sykmeldingsperioder: List<Periode>,
+    val agp: Arbeidsgiverperiode?,
+    val inntekt: Inntekt,
+    val refusjon: Refusjon,
+    val aarsakInnsending: AarsakInnsending,
+)
